refactor(core): Move more code out of NodeExecutionFunctions, and add unit tests (no-changelog) (#13131)
