+++ packages/core/src/execution-engine/__tests__/active-workflows.test.ts
@@ -1,6 +1,5 @@
 import { mock } from 'jest-mock-extended';
 import type {
-	IGetExecuteTriggerFunctions,
 	INode,
 	ITriggerResponse,
 	IWorkflowExecuteAdditionalData,
@@ -15,6 +14,7 @@ import { LoggerProxy, TriggerCloseError, WorkflowActivationError } from 'n8n-wor
 import type { ErrorReporter } from '@/errors/error-reporter';
 
 import { ActiveWorkflows } from '../active-workflows';
+import type { IGetExecuteTriggerFunctions } from '../interfaces';
 import type { PollContext } from '../node-execution-context';
 import type { ScheduledTaskManager } from '../scheduled-task-manager';
 import type { TriggersAndPollers } from '../triggers-and-pollers';
