+++ packages/core/src/execution-engine/interfaces.ts
@@ -1,0 +1,29 @@
+import type {
+	INode,
+	IPollFunctions,
+	ITriggerFunctions,
+	IWorkflowExecuteAdditionalData,
+	Workflow,
+	WorkflowActivateMode,
+	WorkflowExecuteMode,
+} from 'n8n-workflow';
+
+export interface IGetExecutePollFunctions {
+	(
+		workflow: Workflow,
+		node: INode,
+		additionalData: IWorkflowExecuteAdditionalData,
+		mode: WorkflowExecuteMode,
+		activation: WorkflowActivateMode,
+	): IPollFunctions;
+}
+
+export interface IGetExecuteTriggerFunctions {
+	(
+		workflow: Workflow,
+		node: INode,
+		additionalData: IWorkflowExecuteAdditionalData,
+		mode: WorkflowExecuteMode,
+		activation: WorkflowActivateMode,
+	): ITriggerFunctions;
+}
