+++ packages/core/src/execution-engine/node-execution-context/execute-single-context.ts
@@ -14,16 +14,15 @@ import type {
 import { ApplicationError, createDeferredPromise, NodeConnectionType } from 'n8n-workflow';
 
 // eslint-disable-next-line import/no-cycle
+import { getRequestHelperFunctions, returnJsonArray } from '@/node-execute-functions';
+
+import { BaseExecuteContext } from './base-execute-context';
 import {
 	assertBinaryData,
 	detectBinaryEncoding,
 	getBinaryDataBuffer,
 	getBinaryHelperFunctions,
-	getRequestHelperFunctions,
-	returnJsonArray,
-} from '@/node-execute-functions';
-
-import { BaseExecuteContext } from './base-execute-context';
+} from './utils/binary-helper-functions';
 
 export class ExecuteSingleContext extends BaseExecuteContext implements IExecuteSingleFunctions {
 	readonly helpers: IExecuteSingleFunctions['helpers'];
