+++ packages/core/src/execution-engine/node-execution-context/index.ts
@@ -13,3 +13,6 @@ export { TriggerContext } from './trigger-context';
 export { WebhookContext } from './webhook-context';
 
 export { getAdditionalKeys } from './utils/get-additional-keys';
+export { parseIncomingMessage } from './utils/parse-incoming-message';
+export { parseRequestObject } from './utils/parse-request-object';
+export * from './utils/binary-helper-functions';
