+++ packages/core/src/execution-engine/node-execution-context/poll-context.ts
@@ -11,13 +11,13 @@ import { ApplicationError, createDeferredPromise } from 'n8n-workflow';
 
 // eslint-disable-next-line import/no-cycle
 import {
-	getBinaryHelperFunctions,
 	getRequestHelperFunctions,
 	getSchedulingFunctions,
 	returnJsonArray,
 } from '@/node-execute-functions';
 
 import { NodeExecutionContext } from './node-execution-context';
+import { getBinaryHelperFunctions } from './utils/binary-helper-functions';
 
 const throwOnEmit = () => {
 	throw new ApplicationError('Overwrite PollContext.__emit function');
