+++ packages/core/src/execution-engine/node-execution-context/utils/__tests__/deduplication-helper-functions.test.ts
@@ -1,0 +1,25 @@
+import { mock } from 'jest-mock-extended';
+import type { Workflow, INode } from 'n8n-workflow';
+
+import { getDeduplicationHelperFunctions } from '../deduplication-helper-functions';
+
+describe('getDeduplicationHelperFunctions', () => {
+	const workflow = mock<Workflow>();
+	const node = mock<INode>();
+	const helperFunctions = getDeduplicationHelperFunctions(workflow, node);
+
+	it('should create helper functions with correct context', () => {
+		const expectedMethods = [
+			'checkProcessedAndRecord',
+			'checkProcessedItemsAndRecord',
+			'removeProcessed',
+			'clearAllProcessedItems',
+			'getProcessedDataCount',
+		] as const;
+
+		expectedMethods.forEach((method) => {
+			expect(helperFunctions).toHaveProperty(method);
+			expect(typeof helperFunctions[method]).toBe('function');
+		});
+	});
+});
